Resolve "Multiply photon flux with detector time step in photon generation models"

Closes #325

See merge request esa/pyxel!305

+++ pyxel/data_structure/photon.py
@@ -19,7 +19,7 @@ class Photon(Array):
     """
 
     # TODO: add unit (ph)
-    EXP_TYPE = int
+    EXP_TYPE = float
     TYPE_LIST = (
         np.int32,
         np.int64,
